+++ src/IRecordWriter.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2014-2015, Pacific Biosciences of California, Inc.
+// Copyright (c) 2016, Pacific Biosciences of California, Inc.
 //
 // All rights reserved.
 //
@@ -32,45 +32,17 @@
 // OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
 // OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
 // SUCH DAMAGE.
+//
+// File Description
+/// \file IRecordWriter.cpp
+/// \brief Implements the IRecordWriter class.
+//
+// Author: Derek Barnett
 
-// Author: Lance Hepler
+#include "pbbam/IRecordWriter.h"
+using namespace PacBio;
+using namespace PacBio::BAM;
 
-#include "ChemistryTable.h"
+IRecordWriter::IRecordWriter(void) { }
 
-namespace PacBio {
-namespace BAM {
-namespace internal {
-
-extern const std::vector<std::array<std::string, 4>> ChemistryTable = {
-
-    // BindingKit, SequencingKit, BasecallerVersion, Chemistry
-
-    // RS
-    {{"100356300",   "100356200",   "2.1", "P6-C4"}},
-    {{"100356300",   "100356200",   "2.3", "P6-C4"}},
-    {{"100356300",   "100612400",   "2.1", "P6-C4"}},
-    {{"100356300",   "100612400",   "2.3", "P6-C4"}},
-    {{"100372700",   "100356200",   "2.1", "P6-C4"}},
-    {{"100372700",   "100356200",   "2.3", "P6-C4"}},
-    {{"100372700",   "100612400",   "2.1", "P6-C4"}},
-    {{"100372700",   "100612400",   "2.3", "P6-C4"}},
-
-    // 3.0 ("Dromedary"): S/P1-C1/beta
-    {{"100-619-300", "100-620-000", "3.0", "S/P1-C1/beta"}},
-    {{"100-619-300", "100-620-000", "3.1", "S/P1-C1/beta"}},
-
-    // 3.1 ("Echidna"): S/P1-C1.1
-    {{"100-619-300", "100-867-300", "3.1", "S/P1-C1.1"}},
-    {{"100-619-300", "100-867-300", "3.2", "S/P1-C1.1"}},
-
-
-    // 3.1.1 ("Flea"): S/P1-C1.2
-    {{"100-619-300", "100-902-100", "3.1", "S/P1-C1.2"}},
-    {{"100-619-300", "100-902-100", "3.2", "S/P1-C1.2"}}
-
-
-};
-
-} // namespace internal
-} // namespace BAM
-} // namespace PacBio
+IRecordWriter::~IRecordWriter(void) { }