+++ tests/files.cmake
@@ -39,6 +39,7 @@ set( PacBioBAMTest_CPP
     ${PacBioBAM_TestsDir}/src/test_QualityValues.cpp
     ${PacBioBAM_TestsDir}/src/test_ReadAccuracyQuery.cpp
     ${PacBioBAM_TestsDir}/src/test_ReadGroupInfo.cpp
+    ${PacBioBAM_TestsDir}/src/test_SamWriter.cpp
     ${PacBioBAM_TestsDir}/src/test_SequenceUtils.cpp
     ${PacBioBAM_TestsDir}/src/test_StringUtils.cpp
     ${PacBioBAM_TestsDir}/src/test_SubreadLengthQuery.cpp